+++ indico/web/flask/errors.py
@@ -22,7 +22,7 @@ import re
 from flask import current_app, request, send_from_directory, session
 from itsdangerous import BadData
 from sqlalchemy.exc import DatabaseError
-from werkzeug.exceptions import BadRequest, BadRequestKeyError, Forbidden, HTTPException, NotFound, Unauthorized
+from werkzeug.exceptions import BadRequest, BadRequestKeyError, Forbidden, HTTPException, NotFound
 
 from indico.core.errors import IndicoError, get_error_description
 from indico.core.logger import Logger, sentry_log_exception
@@ -89,12 +89,12 @@ def handle_http_exception(exc):
 
 @errors_bp.app_errorhandler(BadData)
 def handle_baddata(exc):
-    return render_error(exc, _('Invalid or expired token'), exc.message, 400)
+    return render_error(exc, _('Invalid or expired token'), to_unicode(exc.message), 400)
 
 
 @errors_bp.app_errorhandler(IndicoError)
 def handle_indico_exception(exc):
-    return render_error(exc, _('Something went wrong'), exc.message, getattr(exc, 'http_status_code', 500))
+    return render_error(exc, _('Something went wrong'), to_unicode(exc.message), getattr(exc, 'http_status_code', 500))
 
 
 @errors_bp.app_errorhandler(DatabaseError)
@@ -104,11 +104,11 @@ def handle_databaseerror(exc):
 
 @errors_bp.app_errorhandler(Exception)
 def handle_exception(exc, message=None):
-    Logger.get('flask').exception(exc.message or 'Uncaught Exception')
+    Logger.get('flask').exception(to_unicode(exc.message) or 'Uncaught Exception')
     if not current_app.debug or request.is_xhr or request.is_json:
         sentry_log_exception()
         if message is None:
-            message = '{}: {}'.format(type(exc).__name__, exc)
+            message = '{}: {}'.format(type(exc).__name__, to_unicode(exc.message))
         return render_error(exc, _('Something went wrong'), message, 500)
     # Let the exception propagate to middleware /the webserver.
     # This triggers the Flask debugger in development and sentry